fix float bug

+++ cli.py
@@ -136,7 +136,7 @@ def parse_args():
                         help="L1 box coefficient in the matching cost")
     parser.add_argument('--eos_coef', default=0.1, type=float,
                         help="Relative classification weight of the no-object class")
-    parser.add_argument('--fake_batch', default=1, type=float,
+    parser.add_argument('--fake_batch', default=1, type=int,
                         help="Relative classification weight of the no-object class")
 
     parser.add_argument('--max_training_sentences', default=3, type=int)
